Addressed CR comments
